restructured integration tests + tests for vec_deque macros
